+++ packages/ui/CHANGELOG.md
@@ -5,6 +5,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ## [unreleased]
 
 ### Added
+
 - `ui/Layout.Cell` prop `grow` to expand its size to the container
 
 ## [4.6.2] - 2023-03-09