+++ packages/ui/CHANGELOG.md
@@ -10,7 +10,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
-- ui/ViewManager to set index prop properly when reverseTransition prop is given
+- `ui/ViewManager` to set index prop properly when reverseTransition prop is given
 
 ## [4.6.2] - 2023-03-09