+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `ui/Layout.Cell` prop `grow` to expand its size to the container
+
+### Fixed
+
+- `ui/ViewManager` to set index prop properly when reverseTransition prop is given
+
+## [4.6.2] - 2023-03-09
+
+No significant changes.
+
 ## [4.6.1] - 2023-02-03
 
 ### Added